added better graphviz handling
+ moved code to embed pngs as matplotlib images (so they render on github) to new eda.utils.py
+ added code to model.utils.py to call model render to graphiz, and insert dpi in the case of png and save to file
+ combined these let you create and display graph representation in a Notebook on Github which is handy
