+++ oreum_core/model/__init__.py
@@ -1,8 +1,9 @@
 # copyright 2021 Oreum OÜ
 from oreum_core.model.base import BasePYMC3Model
-from oreum_core.model.file_io import (
+from oreum_core.model.utils import (
     read_azid,
-    write_azid
+    write_azid,
+    save_graph,
 )
 from oreum_core.model.calc import (
     calc_binary_performance_measures,
