+++ oreum_core/model/utils.py
@@ -1,9 +1,9 @@
-# model.file_io.py
+# model.utils.py
 # copyright 2021 Oreum OÜ
 import errno
 import os
-# import warnings
 import arviz as az
+import pymc3 as pm
 
 def read_azid(dir_traces=[], fn='azid'):
     """ Convenience: read arviz.InferenceData object from file """
@@ -24,3 +24,24 @@ def write_azid(mdl, dir_traces=[]):
         
     mdl.idata.to_netcdf(fqn)
     return f'Wrote: {fqn}'
+
+
+def save_graph(mdl, fp_as_list=[], format='png'):
+    """ Accept a BasePYMC3Model object mdl, get the graphviz representation,
+        write to file and return the fqn
+    """
+
+    gv = pm.model_graph.model_to_graphviz(mdl.model)
+    fqn = os.path.join(*fp_as_list, f'{mdl.name}')  
+
+    if format == 'png':
+        gv.attr(dpi = '300')
+    elif format == 'svg':
+        pass
+    else:
+        raise AttributeError
+    
+    # auto adds the file extension
+    gv.render(filename=fqn, format=format, cleanup=True) 
+
+    return fqn + f'.{format}'
